Started implementing FETCH.
Still missing: converting file to bytes and writing bytes to output.

+++ EZShare/src/EZShare/Resource.java
@@ -21,7 +21,7 @@ public class Resource extends JsonModel {
 	public String channel;
 	public String owner;
 	public String ezserver;
-	public Integer resourceSize;
+	public Long resourceSize;
 
 	/**
 	 * Default constructor
